fix: caught error types
TypeScript now treats errors as `unknown` which causes type errors when
properties like "message" and "path" are accessed. This is maybe not the
most ideal solution. Ideally specific error instances should be caught
and handled with an additional branch for handling unanticipated errors.
For the scope of this PR though, I wanted to change as little as
possible — basically just get the builds running again while leaving the
actual implementation the same.