chore: Reenable package-lock.json
We really need some kind of lock for this (looks like we're aligning on
npm anyway), since master is currently completely unable to be compiled
due to lots of minor patches that have bitrotted the codebase.

I'm now going to try and fix all the issues so this is a standard,
compiling state to actually work from.