fix: local builds with updated dependencies. (#69)
* fix: locale names in test

It doesn't look like "en" is a valid locale on any OS. I think Windows
uses "en-US" and linux/macos use "en_US.utf8". This was a necessary
change for the tests to pass on my machine (manjaro linux). Not sure how
it will impact tests on other platforms.

* fix: caught error types

TypeScript now treats errors as `unknown` which causes type errors when
properties like "message" and "path" are accessed. This is maybe not the
most ideal solution. Ideally specific error instances should be caught
and handled with an additional branch for handling unanticipated errors.
For the scope of this PR though, I wanted to change as little as
possible — basically just get the builds running again while leaving the
actual implementation the same.

* fix: pin flow-remove-types to patched version