chore: upgrade to read-pkg-up ^8.0.0 (from ^7.0.1)
Upgrade the dependency read-pkg-up to ^8.0.0 in order to transitively
update the dependencies read-pkg, normalize-package-data, and semver. In
particular:
- `read-pkg` got bumped from `^5.2.0` to `^6.0.0`
- `normalize-package-data` got bumped from `^2.5.0` to `^3.0.2`
- `semver` got bumped from `2 || 3 || 4 || 5` to `^7.3.4`

which means this package will be compatible with a version of semver
not vulnerable to CVE-2022-25883 going forward.

I upgrade to 8.0.0 instead of latest (9.0.1 at the time of writing)
because it requires at least Node.js v14.13.1, which is higher than
this package and so incompatible.